--- src/main/java/roomescape/admin/presentation/AdminController.java
@@ -1,66 +1,0 @@
-package roomescape.admin.presentation;
-
-import jakarta.validation.Valid;
-import java.net.URI;
-import java.util.List;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-import roomescape.admin.dto.request.AdminReservationRequest;
-import roomescape.reservation.dto.request.ReservationCondition;
-import roomescape.reservation.dto.response.ReservationResponse;
-import roomescape.reservation.dto.response.WaitingResponse;
-import roomescape.reservation.service.ReservationService;
-import roomescape.reservation.service.WaitingService;
-
-@RestController
-@RequestMapping("/api/admin")
-public class AdminController {
-
-    private final ReservationService reservationService;
-    private final WaitingService waitingService;
-
-    public AdminController(ReservationService reservationService, WaitingService waitingService) {
-        this.reservationService = reservationService;
-        this.waitingService = waitingService;
-    }
-
-    @PostMapping("/reservations")
-    public ResponseEntity<ReservationResponse> createReservation(
-            @Valid @RequestBody AdminReservationRequest request) {
-        ReservationResponse reservationResponse = reservationService.createReservation(
-                request.memberId(), request.timeId(), request.themeId(), request.date()
-        );
-        return ResponseEntity
-                .created(URI.create("/reservations/" + reservationResponse.id()))
-                .body(reservationResponse);
-    }
-
-    @GetMapping("/reservations")
-    public List<ReservationResponse> getReservations(@ModelAttribute ReservationCondition condition) {
-        return reservationService.findReservations(condition);
-    }
-
-    @GetMapping("/waitings")
-    public List<WaitingResponse> getAllWaiting() {
-        return waitingService.findAllWaitings();
-    }
-
-    @PostMapping("/waitings/{waitingId}")
-    public ResponseEntity<Void> convertToReservation(@PathVariable("waitingId") Long waitingId) {
-        waitingService.convertWaitingToReservation(waitingId);
-        return ResponseEntity.noContent().build();
-    }
-
-    @DeleteMapping("/waitings/{waitingId}")
-    public ResponseEntity<Void> denyWaiting(@PathVariable("waitingId") Long waitingId) {
-        waitingService.deleteWaitingById(waitingId);
-        return ResponseEntity.noContent().build();
-    }
-}
