+++ src/main/java/roomescape/reservation/presentation/ReservationController.java
@@ -6,19 +6,20 @@ import java.util.List;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.auth.domain.Authenticated;
+import roomescape.reservation.dto.request.AdminReservationRequest;
+import roomescape.reservation.dto.request.ReservationCondition;
 import roomescape.reservation.dto.request.ReservationRequest;
 import roomescape.reservation.dto.response.ReservationResponse;
 import roomescape.reservation.dto.response.ReservationWithStatusResponse;
 import roomescape.reservation.service.ReservationService;
 
 @RestController
-@RequestMapping(value = "/api/reservations")
 public class ReservationController {
 
     private final ReservationService reservationService;
@@ -27,7 +28,7 @@ public class ReservationController {
         this.reservationService = reservationService;
     }
 
-    @PostMapping()
+    @PostMapping("/api/reservations")
     public ResponseEntity<ReservationResponse> createNewReservation(
             @Authenticated Long memberId,
             @Valid @RequestBody ReservationRequest request) {
@@ -39,14 +40,30 @@ public class ReservationController {
                 .body(reservationResponse);
     }
 
-    @DeleteMapping("/{id}")
+    @PostMapping("/api/admin/reservations")
+    public ResponseEntity<ReservationResponse> createReservation(
+            @Valid @RequestBody AdminReservationRequest request) {
+        ReservationResponse reservationResponse = reservationService.createReservation(
+                request.memberId(), request.timeId(), request.themeId(), request.date()
+        );
+        return ResponseEntity
+                .created(URI.create("/api/reservations/" + reservationResponse.id()))
+                .body(reservationResponse);
+    }
+
+    @GetMapping("/api/admin/reservations")
+    public List<ReservationResponse> getReservations(@ModelAttribute ReservationCondition condition) {
+        return reservationService.findReservations(condition);
+    }
+
+    @GetMapping("/api/reservations/my")
+    public List<ReservationWithStatusResponse> getMyReservation(@Authenticated Long id) {
+        return reservationService.findReservationByMemberId(id);
+    }
+
+    @DeleteMapping("/api/reservations/{id}")
     public ResponseEntity<Void> deleteReservation(@PathVariable Long id) {
         reservationService.cancelReservationAndPromoteWait(id);
         return ResponseEntity.noContent().build();
-    }
-
-    @GetMapping("/my")
-    public List<ReservationWithStatusResponse> getMyReservation(@Authenticated Long id) {
-        return reservationService.findReservationByMemberId(id);
     }
 }
