+++ src/main/java/roomescape/reservation/presentation/WaitingController.java
@@ -9,7 +9,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.auth.domain.Authenticated;
 import roomescape.reservation.dto.request.WaitingRequest;
@@ -18,7 +17,6 @@ import roomescape.reservation.dto.response.WaitingWithRankResponse;
 import roomescape.reservation.service.WaitingService;
 
 @RestController
-@RequestMapping("/api/waiting")
 public class WaitingController {
 
     private final WaitingService waitingService;
@@ -27,19 +25,36 @@ public class WaitingController {
         this.waitingService = waitingService;
     }
 
-    @PostMapping
+    @PostMapping("/api/waiting")
     public ResponseEntity<WaitingResponse> addWaitList(@Authenticated Long memberId,
                                                        @RequestBody @Valid WaitingRequest request) {
         WaitingResponse response = waitingService.createWaiting(memberId, request);
         return ResponseEntity.created(URI.create("/api/waiting/" + response.id())).body(response);
     }
 
-    @GetMapping("/my")
+    @PostMapping("/api/admin/waitings/{waitingId}")
+    public ResponseEntity<Void> convertToReservation(@PathVariable("waitingId") Long waitingId) {
+        waitingService.convertWaitingToReservation(waitingId);
+        return ResponseEntity.noContent().build();
+    }
+
+    @GetMapping("/api/admin/waitings")
+    public List<WaitingResponse> getAllWaiting() {
+        return waitingService.findAllWaitings();
+    }
+
+    @GetMapping("/api/waiting/my")
     public List<WaitingWithRankResponse> getMyWaitings(@Authenticated Long memberId) {
         return waitingService.findWaitingByMemberId(memberId);
     }
 
-    @DeleteMapping("/{waitingId}")
+    @DeleteMapping("/api/admin/waitings/{waitingId}")
+    public ResponseEntity<Void> denyWaiting(@PathVariable("waitingId") Long waitingId) {
+        waitingService.deleteWaitingById(waitingId);
+        return ResponseEntity.noContent().build();
+    }
+
+    @DeleteMapping("/api/waiting/{waitingId}")
     public ResponseEntity<Void> cancelWaiting(@Authenticated Long memberId, @PathVariable("waitingId") Long waitingId) {
         waitingService.deleteWaitingById(memberId, waitingId);
         return ResponseEntity.noContent().build();
