+++ src/test/java/roomescape/api/ReservationApiTest.java
@@ -2,6 +2,7 @@ package roomescape.api;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.hamcrest.Matchers.equalTo;
+import static org.hamcrest.Matchers.is;
 
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
@@ -110,4 +111,131 @@ public class ReservationApiTest {
 
         assertThat(reservationRepository.findById(reservation.getId())).isEmpty();
     }
+
+    @Test
+    void 예약_전체_조회() {
+        Member admin = memberRepository.save(
+                Member.builder()
+                        .name("admin")
+                        .email("admin@domain.com")
+                        .password("password1")
+                        .role(Role.ADMIN).build()
+        );
+        Member member = memberRepository.save(
+                Member.builder()
+                        .name("name1")
+                        .email("email1@domain.com")
+                        .password("password1")
+                        .role(Role.MEMBER).build()
+        );
+        Theme theme = themeRepository.save(
+                Theme.builder()
+                        .name("name")
+                        .description("desc")
+                        .thumbnail("thumb").build()
+        );
+        TimeSlot timeSlot = timeSlotRepository.save(
+                TimeSlot.builder()
+                        .startAt(LocalTime.of(9, 0)).build());
+        reservationRepository.save(
+                Reservation.builder()
+                        .member(member)
+                        .reservationTime(new ReservationTime(LocalDate.of(2025, 1, 1), timeSlot))
+                        .theme(theme).build()
+        );
+        reservationRepository.save(
+                Reservation.builder()
+                        .member(member)
+                        .reservationTime(new ReservationTime(LocalDate.of(2025, 1, 2), timeSlot))
+                        .theme(theme).build()
+        );
+        String token = tokenProvider.createToken(admin.getId().toString(), admin.getRole());
+
+        RestAssured.given().log().all()
+                .cookie("token", token)
+                .when().get("api/admin/reservations?memberId=2&themeId=1&dateFrom=2025-01-01&dateTo=2025-01-02")
+                .then().log().all()
+                .statusCode(200)
+                .body("size()", is(2));
+    }
+
+    @Test
+    void 관리자_예약_추가_성공() {
+        Member admin = memberRepository.save(
+                Member.builder()
+                        .name("admin")
+                        .email("admin@domain.com")
+                        .password("password1")
+                        .role(Role.ADMIN).build()
+        );
+        Member member = memberRepository.save(
+                Member.builder()
+                        .name("member1")
+                        .email("email1@domain.com")
+                        .password("password1")
+                        .role(Role.MEMBER).build()
+        );
+        Theme theme = themeRepository.save(
+                Theme.builder()
+                        .name("theme")
+                        .description("desc")
+                        .thumbnail("thumb").build()
+        );
+        TimeSlot timeSlot = timeSlotRepository.save(
+                TimeSlot.builder()
+                        .startAt(LocalTime.of(9, 0)).build());
+        String token = tokenProvider.createToken(admin.getId().toString(), admin.getRole());
+
+        Map<String, Object> reservation = new HashMap<>();
+        reservation.put("date", "2026-08-05");
+        reservation.put("memberId", member.getId());
+        reservation.put("timeId", theme.getId());
+        reservation.put("themeId", timeSlot.getId());
+
+        RestAssured.given().log().all()
+                .contentType(ContentType.JSON)
+                .body(reservation)
+                .cookie("token", token)
+                .when().post("/api/admin/reservations")
+                .then().log().all()
+                .statusCode(201)
+                .body("date", equalTo("2026-08-05"))
+                .body("memberName", equalTo("member1"));
+    }
+
+    @Test
+    void 관리자가_아닌_사용자가_관리자_예약_추가할_경우_403에러가_발생한다() {
+        Member member = memberRepository.save(
+                Member.builder()
+                        .name("name1")
+                        .email("email1@domain.com")
+                        .password("password1")
+                        .role(Role.MEMBER).build()
+        );
+        Theme theme = themeRepository.save(
+                Theme.builder()
+                        .name("name")
+                        .description("desc")
+                        .thumbnail("thumb").build()
+        );
+        TimeSlot timeSlot = timeSlotRepository.save(
+                TimeSlot.builder()
+                        .startAt(LocalTime.of(9, 0)).build());
+        String token = tokenProvider.createToken(member.getId().toString(), member.getRole());
+
+        Map<String, Object> reservation = new HashMap<>();
+        reservation.put("date", "2026-08-05");
+        reservation.put("memberId", member.getId());
+        reservation.put("timeId", theme.getId());
+        reservation.put("themeId", timeSlot.getId());
+
+        RestAssured.given().log().all()
+                .contentType(ContentType.JSON)
+                .body(reservation)
+                .cookie("token", token)
+                .when().post("/api/admin/reservations")
+                .then().log().all()
+                .statusCode(403);
+    }
+
 }
